requirements.txt
Django==3.2.25
paho-mqtt==1.6.1
django-crispy-forms==1.14.0
